refactor(desktop): 将 skillRootKindLabel 迁至共享模块
新增 skill-rule-labels.ts，消除 rules 面板对 skills 面板的 cross-import。
