+++ apps/desktop/package.json
@@ -30,7 +30,7 @@
     "dist": "npm run build && electron-builder --config electron-builder.yml",
     "dist:dir": "npm run build && electron-builder --config electron-builder.yml --dir",
     "test:host": "npm run build:electron && node --test test/host/*.test.mjs",
-    "test:lib": "node scripts/check-renderer-host-internal-imports.mjs && node --experimental-strip-types --test test/workspace-tool-tabs.test.mjs test/browser-url.test.mjs test/browser-element-picker.test.mjs test/local-listeners-parse.test.mjs test/lib/conversation-continue-ui.test.mjs test/lib/desktop-mica-surface.test.mjs test/lib/desktop-shell.test.mjs test/lib/desktop-notification-copy.test.mjs test/lib/empty-session-greeting.test.mjs test/lib/spirit-notification-protocol.test.mjs test/lib/edit-file-line-delta.test.mjs test/lib/markdown-url-transform.test.mjs test/lib/settings-page-titles.test.mjs test/hooks/use-anchored-item-switch.test.mjs test/hooks/github-device-login-model.test.mjs test/lib/process-seal-animation.test.mjs test/lib/composer-branch-checkout-gate.test.mjs test/lib/message-rewind-eligibility.test.mjs test/lib/desktop-keyboard-shortcut-eligibility.test.mjs test/lib/fork-session-display-name.test.mjs test/lib/fork-session-truncation.test.mjs test/lib/fork-eligibility.test.mjs && npx --yes tsx --test test/composer-segments.test.mjs test/lib/file-snippet-wire-text.test.mjs test/lib/file-snippet-chip-dom.test.mjs test/lib/settings-split.test.mjs test/skill-slash.test.mjs test/lib/file-tool-diff-source.test.mjs test/lib/tool-call-unified-diff.test.mjs test/lib/tool-call-display.test.mjs test/lib/process-tool-category.test.mjs test/lib/process-summary-format.test.mjs test/lib/conversation-process-groups.test.mjs test/lib/message-turn-actions-ui.test.mjs test/lib/message-card-spacing-process.test.mjs test/lib/workspace-explorer-icon.test.mjs test/lib/workspace-entry-path-sync.test.mjs test/lib/action-palette.test.mjs test/lib/windows-toast-xml.test.mjs test/lib/model-catalog-detail.test.mjs test/lib/context-usage.test.mjs test/lib/format-compact-token-count.test.mjs test/lib/model-picker-shortcut-bridge.test.mjs && npm run build:electron && node --test test/conversation-thinking-ui.test.mjs test/lib/git-changes-menu-items.test.mjs",
+    "test:lib": "node scripts/check-renderer-host-internal-imports.mjs && node --experimental-strip-types --test test/workspace-tool-tabs.test.mjs test/browser-url.test.mjs test/browser-element-picker.test.mjs test/local-listeners-parse.test.mjs test/lib/conversation-continue-ui.test.mjs test/lib/desktop-mica-surface.test.mjs test/lib/desktop-shell.test.mjs test/lib/desktop-notification-copy.test.mjs test/lib/empty-session-greeting.test.mjs test/lib/spirit-notification-protocol.test.mjs test/lib/edit-file-line-delta.test.mjs test/lib/markdown-url-transform.test.mjs test/lib/settings-page-titles.test.mjs test/hooks/use-anchored-item-switch.test.mjs test/hooks/github-device-login-model.test.mjs test/lib/process-seal-animation.test.mjs test/lib/composer-branch-checkout-gate.test.mjs test/lib/message-rewind-eligibility.test.mjs test/lib/desktop-keyboard-shortcut-eligibility.test.mjs test/lib/fork-session-display-name.test.mjs test/lib/fork-session-truncation.test.mjs test/lib/fork-eligibility.test.mjs && npx --yes tsx --test test/composer-segments.test.mjs test/lib/file-snippet-wire-text.test.mjs test/lib/file-snippet-chip-dom.test.mjs test/lib/settings-split.test.mjs test/lib/skill-rule-labels.test.mjs test/skill-slash.test.mjs test/lib/file-tool-diff-source.test.mjs test/lib/tool-call-unified-diff.test.mjs test/lib/tool-call-display.test.mjs test/lib/process-tool-category.test.mjs test/lib/process-summary-format.test.mjs test/lib/conversation-process-groups.test.mjs test/lib/message-turn-actions-ui.test.mjs test/lib/message-card-spacing-process.test.mjs test/lib/workspace-explorer-icon.test.mjs test/lib/workspace-entry-path-sync.test.mjs test/lib/action-palette.test.mjs test/lib/windows-toast-xml.test.mjs test/lib/model-catalog-detail.test.mjs test/lib/context-usage.test.mjs test/lib/format-compact-token-count.test.mjs test/lib/model-picker-shortcut-bridge.test.mjs && npm run build:electron && node --test test/conversation-thinking-ui.test.mjs test/lib/git-changes-menu-items.test.mjs",
     "preview": "vite preview --host 127.0.0.1 --port 4173",
     "start": "electron dist-electron/electron/main.js",
     "gen:icons": "node scripts/gen-app-icons.mjs",
